+++ Code/Graphic/GUI/WindowRenderer.cpp
@@ -1,7 +1,7 @@
 #include "WindowRenderer.h"
 
 int WindowRenderer::render(ska::Window*, ska::Texture* tex, int x, int y, ska::Rectangle* clip) const{
-	ska::Rectangle destBuf = { x, y, tex->getWidth(), tex->getHeight() };
+	ska::Rectangle destBuf = { x, y, static_cast<int>(tex->getWidth()), static_cast<int>(tex->getHeight()) };
 
 	if (clip != nullptr) {
 		destBuf.w = clip->w;
