+++ Code/Script/Commands/CommandExpulse.cpp
@@ -41,7 +41,7 @@ std::string CommandExpulse::execute(ska::ScriptComponent& script, std::vector<st
 	const ska::HitboxComponent& hc = m_entityManager.getComponent<ska::HitboxComponent>(internalEntity);
 	const ska::Point<int>& centerPos = ska::PositionComponent::getCenterPosition(pc, hc);
 
-	ska::Rectangle hitbox{ pc.x + hc.xOffset, pc.y + hc.yOffset, hc.width, hc.height };
+	ska::Rectangle hitbox{ pc.x + hc.xOffset, pc.y + hc.yOffset, static_cast<int>(hc.width), static_cast<int>(hc.height) };
 	const ska::Point<int> offsetHitboxBlock = m_world.alignOnBlock(hitbox);
 	const ska::Rectangle targetBlock = m_world.placeOnNearestPracticableBlock(hitbox, 1);
 	ska::Point<int> vector;
